feat: add select backup screen
